Module subspace_fraud_proof::invalid_state_transition_proof

Invalid state transition proof

This module provides the feature of generating and verifying the execution proof used in the Subspace fraud proof mechanism. The execution is more fine-grained than the entire block execution, block execution hooks (initialize_block and finalize_block) and any specific extrinsic execution are supported.

Structs

CoreDomainExtrinsicsBuilder

Utility to build the core domain extrinsics.

ExecutionProver

Creates storage proof for verifying an execution without owning the whole state.

InvalidStateTransitionProofVerifier

Invalid state transition proof verifier.

PrePostStateRootVerifier

Verifier of pre_state_root in [InvalidStateTransitionProof].

SystemDomainExtrinsicsBuilder

Utility to build the system domain extrinsics.

Traits

BuildDomainExtrinsics

Trait to build the extrinsics of domain block derived from the original primary block.

VerifyInvalidStateTransitionProof

Verifies invalid state transition proof.

VerifyPrePostStateRoot

Verifies pre_state_root and post_state_root in [InvalidStateTransitionProof].